refactor: flatten core/ modules into the top-level pyjinhx package
