+++ tests/21_component_template_tag_expansion_test.py
@@ -4,7 +4,7 @@ import tempfile
 from jinja2 import Environment, FileSystemLoader
 
 from pyjinhx import BaseComponent
-from pyjinhx.core.renderer import Renderer
+from pyjinhx.renderer import Renderer
 
 
 def test_component_template_can_expand_custom_tags():
