+++ tests/27_autodiscovery_test.py
@@ -19,8 +19,8 @@ import pytest
 from jinja2 import Environment, FileSystemLoader
 
 from pyjinhx import Registry
-from pyjinhx.core.tags import ComponentAutodiscover
-from pyjinhx.core.renderer import Renderer
+from pyjinhx.tags import ComponentAutodiscover
+from pyjinhx.renderer import Renderer
 
 _COMPONENT_SRC = """\
 from pyjinhx import BaseComponent
